+++ src/sys/os/browser/web/event/key/key/key.rs
@@ -4,7 +4,6 @@
 //
 
 use crate::{EventKey, EventKind};
-#[cfg(feature = "time")]
 use crate::{EventKindTimed, EventTimestamp, JsInstant, Timed, is};
 use crate::{KeyFfi, KeyMods, KeyState};
 crate::_use! { basic::from_utf8 }
@@ -13,12 +12,10 @@ crate::_use! { basic::from_utf8 }
 /// A web API keyboard event.
 #[doc = crate::_doc_meta!{
     location("sys/os/browser/web"),
-    #[cfg(all(feature = "time", target_pointer_width = "64"))]
+    #[cfg(target_pointer_width = "64")]
     test_size_of(WebEventKey = 32|256; niche Option),
-    #[cfg(all(feature = "time", target_pointer_width = "32"))]
+    #[cfg(target_pointer_width = "32")]
     test_size_of(WebEventKey = 28|224; niche Option),
-    #[cfg(not(feature = "time"))]
-    test_size_of(WebEventKey = 20|160; niche Option),
 }]
 ///
 /// Represents a JavaScript `KeyboardEvent` in a callback-friendly form.
@@ -45,7 +42,6 @@ pub struct WebEventKey {
     pub state: KeyState, // 1 byte
 
     /// The JavaScript event timestamp.
-    #[cfg(feature = "time")]
     pub timestamp: JsInstant, // 8 bytes
 }
 
@@ -56,16 +52,9 @@ impl WebEventKey {
         physical: KeyFfi,
         mods: KeyMods,
         state: KeyState,
-        #[cfg(feature = "time")] timestamp: JsInstant,
+        timestamp: JsInstant,
     ) -> Self {
-        Self {
-            semantic,
-            physical,
-            mods,
-            state,
-            #[cfg(feature = "time")]
-            timestamp,
-        }
+        Self { semantic, physical, mods, state, timestamp }
     }
 }
 impl WebEventKey {
@@ -84,13 +73,10 @@ impl WebEventKey {
             physical: from.physical.to_ffi(),
             mods: from.mods,
             state: from.state,
-            #[cfg(feature = "time")]
             timestamp: JsInstant::ZERO,
         }
     }
 }
-#[cfg(feature = "time")]
-#[cfg_attr(nightly_doc, doc(cfg(feature = "time")))]
 impl WebEventKey {
     /// Converts `WebEventKey` to `EventKindTimed`.
     pub const fn to_event_kind_timed(self) -> EventKindTimed {
@@ -114,15 +100,11 @@ impl From<WebEventKey> for EventKind {
         from.to_event_kind()
     }
 }
-#[cfg(feature = "time")]
-#[cfg_attr(nightly_doc, doc(cfg(feature = "time")))]
 impl From<WebEventKey> for EventKindTimed {
     fn from(from: WebEventKey) -> Self {
         from.to_event_kind_timed()
     }
 }
-#[cfg(feature = "time")]
-#[cfg_attr(nightly_doc, doc(cfg(feature = "time")))]
 impl From<Timed<EventKey, Option<EventTimestamp>>> for WebEventKey {
     fn from(from: Timed<EventKey, Option<EventTimestamp>>) -> Self {
         Self::from_event_key_timed(from)