+++ src/ui/event/event.rs
@@ -3,12 +3,10 @@
 //! Defines [`Event`].
 //
 
-#[cfg(feature = "time")]
-use crate::EventKindTimed;
 use crate::{_impl_init, NonZeroU64};
 use crate::{
-    DeviceId, EventKey, EventKind, EventMouse, EventPointer, EventTag, EventTagSet, EventTarget,
-    EventTimestamp, EventWheel, EventWindow, WindowId,
+    DeviceId, EventKey, EventKind, EventKindTimed, EventMouse, EventPointer, EventTag, EventTagSet,
+    EventTarget, EventTimestamp, EventWheel, EventWindow, WindowId,
 };
 
 #[doc = crate::_tags!(event)]
@@ -118,7 +116,6 @@ impl Event {
     ///
     /// `processed` and `count` are left unset and should be filled by the engine.
     #[inline(always)]
-    #[cfg(feature = "time")]
     pub fn from_kind_timed_with(target: EventTarget, kind: EventKindTimed) -> Event {
         Self { kind: kind.value, target, emitted: kind.time, processed: None, count: None }
     }
@@ -268,7 +265,6 @@ impl From<EventKind> for Event {
         }
     }
 }
-#[cfg(feature = "time")]
 impl From<EventKindTimed> for Event {
     fn from(kind: EventKindTimed) -> Event {
         Self {